Split Sidebar.py into parts/ subfolder
