+++ src/windows/mainWindow/elements/Sidebar/Sidebar.py
@@ -12,37 +12,44 @@ This programm comes with ABSOLUTELY NO WARRANTY!
 You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 """
-# Import gtk modules
-import os
 import gi
 
-from src.backend.DeckManagement.InputIdentifier import Input, InputIdentifier
+import globals as gl
+from GtkHelper.GtkHelper import ErrorPage
 from src.backend.DeckManagement.DeckController import ControllerKey, DeckController
+from src.backend.DeckManagement.InputIdentifier import Input, InputIdentifier
 from src.windows.mainWindow.elements.PageSelector import PageSelector
+from src.windows.mainWindow.elements.Sidebar.elements.ActionChooser import ActionChooser
+from src.windows.mainWindow.elements.Sidebar.elements.ActionConfigurator import ActionConfigurator
 from src.windows.mainWindow.elements.Sidebar.elements.DialEditor import DialEditor
-from src.windows.mainWindow.elements.Sidebar.elements.StateSwitcher import StateSwitcher
 from src.windows.mainWindow.elements.Sidebar.elements.ScreenEditor import ScreenEditor
-
+from src.windows.mainWindow.elements.Sidebar.Parts.KeyEditor import (
+    KeyEditor,
+    KeyEditorKeyBox,
+    TestStack,
+)
+from src.windows.mainWindow.elements.Sidebar.Parts.PageEditor import PageEditor
+from src.windows.mainWindow.elements.Sidebar.Parts.Pages import (
+    AdwPageRow,
+    PageRow,
+    PagesGroup,
+)
 
 gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
-from gi.repository import Gtk, Adw
+from gi.repository import Adw, Gtk
 
-# Import Python modules
-from loguru import logger as log
+__all__ = [
+    "AdwPageRow",
+    "KeyEditor",
+    "KeyEditorKeyBox",
+    "PageEditor",
+    "PageRow",
+    "PagesGroup",
+    "Sidebar",
+    "TestStack",
+]
 
-# Import own modules
-from src.windows.mainWindow.elements.Sidebar.elements.IconSelector import IconSelector
-from src.windows.mainWindow.elements.Sidebar.elements.LabelEditor import LabelEditor
-from src.windows.mainWindow.elements.Sidebar.elements.ActionManager import ActionManager
-from src.windows.mainWindow.elements.Sidebar.elements.ActionChooser import ActionChooser
-from src.windows.mainWindow.elements.Sidebar.elements.ActionConfigurator import ActionConfigurator
-from src.windows.mainWindow.elements.Sidebar.elements.BackgroundEditor import BackgroundEditor
-from src.windows.mainWindow.elements.Sidebar.elements.ImageEditor import ImageEditor
-from GtkHelper.GtkHelper import BetterPreferencesGroup, ErrorPage
-
-# Import globals
-import globals as gl
 
 class Sidebar(Adw.NavigationPage):
     def __init__(self, main_window, **kwargs):
@@ -204,309 +211,3 @@ class Sidebar(Adw.NavigationPage):
 
     def update(self):
         self.load_for_identifier(self.active_identifier, self.active_state)
-
-
-class KeyEditor(Gtk.Box):
-    def __init__(self, sidebar: Sidebar, **kwargs):
-        self.sidebar:Sidebar = sidebar
-        super().__init__(**kwargs)
-        self.set_orientation(Gtk.Orientation.VERTICAL)
-
-        self.scrolled_window = Gtk.ScrolledWindow(hexpand=True, vexpand=True)
-        self.append(self.scrolled_window)
-
-        self.main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, hexpand=True, vexpand=True)
-        self.scrolled_window.set_child(self.main_box)
-
-        self.state_switcher = StateSwitcher("keys", margin_start=20, margin_end=20, margin_top=10, margin_bottom=10, hexpand=True)
-        self.state_switcher.add_switch_callback(self.on_state_switch)
-        self.state_switcher.add_add_new_callback(self.on_add_new_state)
-        self.state_switcher.set_n_states(0)
-        self.main_box.append(self.state_switcher)
-
-        self.icon_selector = IconSelector(sidebar, halign=Gtk.Align.CENTER, margin_top=30)
-        self.main_box.append(self.icon_selector)
-
-        self.image_editor = ImageEditor(sidebar, margin_top=90)
-        self.main_box.append(self.image_editor)
-
-        self.background_editor = BackgroundEditor(sidebar, margin_top=25)
-        self.main_box.append(self.background_editor)
-
-        self.label_editor = LabelEditor(sidebar, margin_top=25)
-        self.main_box.append(self.label_editor)
-
-        self.action_editor = ActionManager(sidebar, margin_top=25, width_request=400)
-        self.main_box.append(self.action_editor)
-
-        self.remove_state_button = Gtk.Button(label="Remove State", css_classes=["destructive-action"], margin_top=15, margin_bottom=15, margin_start=15, margin_end=15)
-        self.remove_state_button.connect("clicked", self.on_remove_state)
-        self.append(self.remove_state_button)
-
-    def on_state_switch(self, *args):
-        state = self.state_switcher.get_selected_state()
-
-        controller = gl.app.main_win.get_active_controller()
-        if controller is None:
-            return
-        
-        controller_input = controller.get_input(self.sidebar.active_identifier)
-        log.info(f"Going to state {state} from {controller_input.state}")
-        controller_input.set_state(state=state, update_sidebar=True)
-
-    def on_add_new_state(self, state):
-        controller = gl.app.main_win.get_active_controller()
-        if controller is None:
-            return
-        
-        c_input = controller.get_input(self.sidebar.active_identifier)
-        c_input.add_new_state()
-
-        self.remove_state_button.set_visible(self.state_switcher.get_n_states() > 1)
-
-    def on_remove_state(self, button):
-        if self.state_switcher.get_n_states() <= 1:
-            return
-
-        controller = gl.app.main_win.get_active_controller()
-        if controller is None:
-            return
-        
-        active_state = self.state_switcher.get_selected_state()
-        
-        c_input = controller.get_input(self.sidebar.active_identifier)
-        c_input.remove_state(active_state)
-
-        self.remove_state_button.set_visible(self.state_switcher.get_n_states() > 1)
-
-    def load_for_identifier(self, identifier: InputIdentifier, state: int):
-        self.sidebar.active_identifier = identifier
-
-        controller = gl.app.main_win.get_active_controller()
-        if controller is None:
-            return
-        
-        c_input = controller.get_input(identifier)
-
-        self.state_switcher.load_for_identifier(identifier, state)
-        c_input.set_state(state)
-
-        self.remove_state_button.set_visible(self.state_switcher.get_n_states() > 1)
-
-        self.icon_selector.load_for_identifier(identifier, state)
-        self.image_editor.load_for_identifier(identifier, state)
-        self.label_editor.load_for_identifier(identifier, state)
-        self.action_editor.load_for_identifier(identifier, state)
-        self.background_editor.load_for_identifier(identifier, state)
-
-class PageEditor(Gtk.Box):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.clamp = Adw.Clamp()
-        self.set_margin_top(40)
-        self.append(self.clamp)
-
-        self.main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, hexpand=True)
-        self.clamp.set_child(self.main_box)
-
-        self.pages_group = PagesGroup()
-        self.main_box.append(self.pages_group)
-
-
-class PagesGroup(BetterPreferencesGroup):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.load_pages()
-
-    def load_pages(self):
-        pages = gl.page_manager.get_pages()
-        for page_path in pages:
-            if os.path.dirname(page_path) != os.path.join(gl.DATA_PATH, "pages"):
-                continue
-            page_row = AdwPageRow(pages_group=self, page_path=page_path)
-            self.add(page_row)
-
-        row1 = Adw.EntryRow(title="Title")
-        self.add(row1)
-
-        child:Gtk.Box = row1.get_child() # Box
-        prefix_box:Gtk.Box = child.get_first_child()
-        gizmo = prefix_box.get_next_sibling()
-        empty_title = gizmo.get_first_child()
-        title = empty_title.get_next_sibling()
-
-        empty_title.set_visible(False)
-        title.set_visible(False)
-
-        row1.add_suffix(Gtk.Button(icon_name="view-more-symbolic", css_classes=["flat"], vexpand=False, valign=Gtk.Align.CENTER))
-
-
-class AdwPageRow(Adw.PreferencesRow):
-    def __init__(self, pages_group:PagesGroup, page_path:str = None):
-        super().__init__(overflow=Gtk.Overflow.HIDDEN, css_classes=["no-padding"])
-
-        self.pages_group = pages_group
-        self.page_path = page_path
-
-        # self.toggle_button = Gtk.ToggleButton(hexpand=True, vexpand=True, css_classes=["no-rounded-corners", "flat"])
-        # self.toggle_button.connect("toggled", self.on_toggled)
-        # self.set_child(self.toggle_button)
-
-        self.main_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, hexpand=True, margin_bottom=6, margin_top=6,
-                                margin_start=10, margin_end=10)
-        self.set_child(self.main_box)
-        # self.toggle_button.set_child(self.main_box)
-
-        self.label = Gtk.Label(xalign=0, label=os.path.splitext(os.path.basename(page_path))[0], hexpand=False,
-                               visible=True, margin_start=2)
-        self.main_box.append(self.label)
-
-        self.entry = Gtk.Entry(text=os.path.splitext(os.path.basename(page_path))[0], hexpand=False, xalign=0,
-                               css_classes=["flat", "no-border", "no-outline"], has_frame=False,
-                               visible=False)
-        self.main_box.append(self.entry)
-
-        self.active_icon = Gtk.Image(icon_name="selection-mode-symbolic", css_classes=["flat"], margin_start=3, visible=False)
-        self.main_box.append(self.active_icon)
-
-        self.edit_button = Gtk.Button(icon_name="document-edit-symbolic", halign=Gtk.Align.END, css_classes=["flat"], hexpand=True)
-        self.edit_button.connect("clicked", self.on_edit_clicked)
-        self.main_box.append(self.edit_button)
-
-        # Click ctrl
-        self.click_ctrl = Gtk.GestureClick.new()
-        self.click_ctrl.set_button(1)
-        self.click_ctrl.connect("pressed", self.on_click)
-        self.main_box.add_controller(self.click_ctrl)
-
-        # Focus ctrl
-        self.focus_ctrl = Gtk.EventControllerFocus()
-        self.main_box.add_controller(self.focus_ctrl)
-        # self.focus_ctrl.connect("enter", self.on_focus_in)
-        self.focus_ctrl.connect("leave", self.on_focus_out)
-
-    def on_toggled(self, button):
-        self.active_icon.set_visible(button.get_active())
-
-    def on_loose_focus(self, *args):
-        self.set_active(False)
-        self.remove_css_class("active-border")
-        self.label.set_visible(True)
-        self.entry.set_visible(False)
-        self.entry.set_hexpand(False)
-        self.active_icon.set_visible(False)
-        self.active_icon.set_hexpand(True)
-
-
-    def on_focus_out(self, *args):
-        return
-        self.on_loose_focus()
-
-    def remove_focus_from_other_pages(self):
-        page_rows = self.pages_group.get_rows()
-        for row in page_rows:
-            if row != self:
-                row.on_loose_focus()
-
-    def on_click(self, gesture, n_press, x, y):
-        self.remove_focus_from_other_pages()
-        if n_press == 1:
-            self.set_active(True)
-            return
-
-        elif n_press == 2:
-            self.on_edit_clicked(self)
-
-    def on_edit_clicked(self, button):
-        self.add_css_class("active-border")
-        show_label = not self.label.get_visible()
-        self.label.set_visible(show_label)
-        self.entry.set_visible(not show_label)
-        self.entry.grab_focus_without_selecting()
-        self.entry.set_position(-1)
-        self.active_icon.set_hexpand(False)
-        self.entry.set_hexpand(True)
-
-    def set_active(self, active: bool):
-        if active:
-            self.set_other_rows_inactive()
-
-        self.active_icon.set_visible(active)
-        if active:
-            self.add_css_class("page-row-active")
-        else:
-            self.remove_css_class("page-row-active")
-
-    def set_other_rows_inactive(self):
-        page_rows = self.pages_group.get_rows()
-        for row in page_rows:
-            if row != self:
-                if hasattr(row, "set_active"):
-                    row.set_active(False)
-        
-
-
-
-
-class PageRow(Gtk.Overlay):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.set_halign(Gtk.Align.CENTER)
-        self.set_size_request(300, -1)
-        self.build()
-
-    def build(self):
-        self.toggle_button = Gtk.ToggleButton(hexpand=False, css_classes=["flat"])
-        self.set_child(self.toggle_button)
-
-        self.label = Gtk.Label(xalign=0, label="Page Row")
-        self.toggle_button.set_child(self.label)
-
-        self.menu_button = Gtk.Button(icon_name="view-more-symbolic", halign=Gtk.Align.END, css_classes=["flat"])
-        self.add_overlay(self.menu_button)
-
-
-class KeyEditorKeyBox(Gtk.Box):
-    def __init__(self, sidebar: Sidebar, **kwargs):
-        super().__init__(**kwargs)
-        self.sidebar:Sidebar = sidebar
-
-        self.scrolled_window = Gtk.ScrolledWindow(hexpand=True, vexpand=True)
-        self.append(self.scrolled_window)
-
-        self.main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, hexpand=True, vexpand=True)
-        self.scrolled_window.set_child(self.main_box)
-
-        self.icon_selector = IconSelector(sidebar, halign=Gtk.Align.CENTER, margin_top=75)
-        self.main_box.append(self.icon_selector)
-
-        self.image_editor = ImageEditor(sidebar, margin_top=25)
-        self.main_box.append(self.image_editor)
-
-        self.background_editor = BackgroundEditor(sidebar, margin_top=25)
-        self.main_box.append(self.background_editor)
-
-        self.label_editor = LabelEditor(sidebar, margin_top=25)
-        self.main_box.append(self.label_editor)
-
-        self.action_editor = ActionManager(sidebar, margin_top=25, width_request=400)
-        self.main_box.append(self.action_editor)
-
-    def load_for_key(self, key: Input.Key, state: int):
-        if not isinstance(key, Input.Key):
-            raise TypeError("Input.Key expected")
-        self.sidebar.active_identifier = key
-        self.sidebar.active_state = state
-        #TODO: Migrate to identifier
-        self.icon_selector.load_for_identifier(key, state)
-        self.image_editor.load_for_identifier(key.coords, state)
-        self.label_editor.load_for_identifier(key, state)
-        self.action_editor.load_for_coords(key.coords, state)
-        self.background_editor.load_for_identifier(key, state)
-
-class TestStack(Gtk.Stack):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-        self.add_titled(Gtk.Label(), "key", "Key")
-        self.add_titled(Gtk.Label(), "pages", "Pages")
